backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pillow
pytesseract
google-cloud-vision==3.4.0
openai
python-multipart
cryptography
pyjwt
fuzzywuzzy
python-Levenshtein
